backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15.0,<3.16.0
django-cors-headers>=4.3.1,<5.0
drf-yasg>=1.21.7,<2.0
